Memoize some internal methods results in order to optimize performance. Per single instance the benefit is quite small, but, when is necessary to operate the big amount of instances - the benefit is significant. I've created benchmark and here is the result before optimization:
```
dmytro@DESKTOP-GHVMRBF:~/work/api.premiervirtual.com$ ./development php artisan benchmark:aggregate-vs-eloquent --batchSize=1000 --limit=10000
 10000/10000 [▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓] 100%
Execution time: 6.1068408489227 seconds
Consumed RAM: 36Mb
dmytro@DESKTOP-GHVMRBF:~/work/api.premiervirtual.com$ ./development php artisan benchmark:aggregate-vs-eloquent --batchSize=1000 --limit=10000
 10000/10000 [▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓] 100%
Execution time: 5.9404318332672 seconds
Consumed RAM: 36Mb
dmytro@DESKTOP-GHVMRBF:~/work/api.premiervirtual.com$ ./development php artisan benchmark:aggregate-vs-eloquent --batchSize=1000 --limit=10000
 10000/10000 [▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓] 100%
Execution time: 6.0607869625092 seconds
Consumed RAM: 36Mb
```

And this is result after optimization:

```
dmytro@DESKTOP-GHVMRBF:~/work/api.premiervirtual.com$ ./development php artisan benchmark:aggregate-vs-eloquent --batchSize=1000 --limit=10000
 10000/10000 [▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓] 100%
Execution time: 3.582909822464 seconds
Consumed RAM: 36Mb
dmytro@DESKTOP-GHVMRBF:~/work/api.premiervirtual.com$ ./development php artisan benchmark:aggregate-vs-eloquent --batchSize=1000 --limit=10000
 10000/10000 [▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓] 100%
Execution time: 3.6363780498505 seconds
Consumed RAM: 36Mb
dmytro@DESKTOP-GHVMRBF:~/work/api.premiervirtual.com$ ./development php artisan benchmark:aggregate-vs-eloquent --batchSize=1000 --limit=10000
 10000/10000 [▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓] 100%
Execution time: 3.5810899734497 seconds
Consumed RAM: 36Mb
dmytro@DESKTOP-GHVMRBF:~/work/api.premiervirtual.com$
```