@segoddnja 2021-11-25
@othercodes Definitely there is something wrong with codecov. And last time there was exactly the same issue. Maybe let's disable this check? Would like to release these optimizations asap.